# Fix variable usage impacting schema build performance